Revert "Update Game.java"
This reverts commit 6a79f3fd89b1543b8287c095574ea4f9a15bafb5.

+++ GUI/Component/Game.java
@@ -5,7 +5,7 @@ import java.util.*;
 
 public class Game implements KeyListener
 {
-    public static double CELL_SIDE = 20;
+    public static int CELL_SIDE = 20;
     Player[] players;
     Map map;
